Improve post CRUD controller

+++ src/Controller/PostController.php
@@ -4,10 +4,10 @@ namespace App\Controller;
 
 use App\Entity\Comment;
 use App\Entity\Post;
-use App\Repository\PostRepository;
 use App\Entity\User;
 use App\Form\CommentType;
 use App\Form\PostType;
+use App\Repository\PostRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -21,67 +21,82 @@ final class PostController extends AbstractController
     {
     }
 
-    #[Route('/post', name: 'app_post')]
+    #[Route('/post', name: 'app_post', methods: [Request::METHOD_GET])]
     public function index(): Response
     {
-        /** @var User */
         $user = $this->getUser();
-        $profile = $user->getProfile();
-        $allPosts = $this->postRepository->getPostsByProfile($profile);
+        if (!$user instanceof User || $user->getProfile() === null) {
+            throw $this->createAccessDeniedException();
+        }
+
         return $this->render('post/index.html.twig', [
-            'posts' => $allPosts,
+            'posts' => $this->postRepository->getPostsByProfile($user->getProfile()),
         ]);
     }
 
     #[Route('/post/create', name: 'app_post_new', methods: [Request::METHOD_GET, Request::METHOD_POST])]
-    public function createPost(Request $request): Response
+    public function create(Request $request): Response
     {
-        $post = new Post();
-        $form = $this->createForm(PostType::class, $post);
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid())
-        {
-            /** @var User */
-            $user = $this->getUser();
-            $profile = $user->getProfile();
-            $post->setProfile($profile);
-
-            $this->postRepository->savePost($post);
-            return $this->redirectToRoute('app_post_show', ['id' => $post->getId()]);
+        $user = $this->getUser();
+        if (!$user instanceof User || $user->getProfile() === null) {
+            throw $this->createAccessDeniedException();
         }
 
-        return $this->render('post/new.html.twig', ['form' => $form]);
+        $post = new Post();
+        $form = $this->createForm(PostType::class, $post);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $post->setProfile($user->getProfile());
+            $this->postRepository->savePost($post);
+
+            return $this->redirectToRoute('app_post_show', ['id' => $post->getId()], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->render('post/new.html.twig', [
+            'form' => $form,
+        ]);
     }
 
-    #[Route('/post/{id}/show', name: 'app_post_show', methods: [Request::METHOD_GET, ])]
-    public function showPost(Post $post): Response
+    #[Route('/post/{id}/show', name: 'app_post_show', methods: [Request::METHOD_GET])]
+    public function show(Post $post): Response
     {
         $comment = new Comment();
-        $commentForm = $this->createForm(CommentType::class, $comment, ['action' => $this->generateUrl('app_comment_new', ['post_id' => $post->getId()])] );
+        $commentForm = $this->createForm(CommentType::class, $comment, [
+            'action' => $this->generateUrl('app_comment_new', ['post_id' => $post->getId()]),
+        ]);
 
-        return $this->render('post/show.html.twig', ['post' => $post, 'form' => $commentForm]);
+        return $this->render('post/show.html.twig', [
+            'post' => $post,
+            'form' => $commentForm,
+        ]);
     }
 
     #[Route('/post/{id}/edit', name: 'app_post_edit', methods: [Request::METHOD_GET, Request::METHOD_POST])]
-    public function editPost(Post $post, Request $request): Response
+    public function edit(Post $post, Request $request): Response
     {
         $form = $this->createForm(PostType::class, $post);
-
         $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) 
-        {
+
+        if ($form->isSubmitted() && $form->isValid()) {
             $this->postRepository->savePost($post);
-            return $this->redirectToRoute('app_post_show', ['id' => $post->getId()]);
+
+            return $this->redirectToRoute('app_post_show', ['id' => $post->getId()], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('post/edit.html.twig', ['form' => $form]);
+        return $this->render('post/edit.html.twig', [
+            'post' => $post,
+            'form' => $form,
+        ]);
     }
 
     #[Route('/post/{id}/delete', name: 'app_post_delete', methods: [Request::METHOD_POST])]
-    public function deletePost(Post $post): Response
+    public function delete(Post $post, Request $request): Response
     {
-        $this->postRepository->deletePost($post);
-        return $this->redirectToRoute('app_post');
+        if ($this->isCsrfTokenValid('delete-post'.$post->getId(), $request->getPayload()->getString('_token'))) {
+            $this->postRepository->deletePost($post);
+        }
+
+        return $this->redirectToRoute('app_post', [], Response::HTTP_SEE_OTHER);
     }
 }
